Memoize the props value with useMemo
The props value that is primitive is actually fine as long as we do not
change its value, the component memoization will work fine.

But, when the value of props is not a primitive, like an object, we must
use useMemo to memoize that object. Remember that in JavaScript, `{}` is
not the same as `{}` even though they are the same empty objects. They have
different references in memory. So JavaScript will think they are different
which automatically assumes they are always a new value to the component.