Memoizing the object value of PostContext.Provider
When the state in the App component is changed, it causes re-render to
App and its children components. Because of this, the Context Provider
component is also be re-rendered, and its value which is an object will
be re-created. So, the component that consumes this context will be
re-rendered as well.

To prevent that, the object value of the Context Provider needs to be
memoized.

After memoizing, the Context Provider will not cause re-render as long
as its object value doesn't change.